backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
httpx==0.25.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2
aiohttp==3.9.1